requirements.txt
Werkzeug
gunicorn
Flask
google-cloud-aiplatform
google-cloud-vision
google-cloud-storage
#appengine-python-standard
google-genai